fix test

+++ testkit/src/main/java/jet/isur/nsi/testkit/test/BaseSqlTest.java
@@ -1,5 +1,30 @@
 package jet.isur.nsi.testkit.test;
 
+import jet.isur.nsi.api.data.DictRow;
+import jet.isur.nsi.api.data.DictRowBuilder;
+import jet.isur.nsi.api.data.NsiConfig;
+import jet.isur.nsi.api.data.NsiConfigAttr;
+import jet.isur.nsi.api.data.NsiConfigDict;
+import jet.isur.nsi.api.data.NsiConfigField;
+import jet.isur.nsi.api.data.NsiQuery;
+import jet.isur.nsi.api.data.NsiQueryAttr;
+import jet.isur.nsi.api.model.MetaAttrType;
+import jet.isur.nsi.common.data.DictDependencyGraph;
+import jet.isur.nsi.common.sql.DefaultSqlDao;
+import jet.isur.nsi.common.sql.DefaultSqlGen;
+import jet.isur.nsi.testkit.utils.DaoUtils;
+import org.jooq.DeleteWhereStep;
+import org.jooq.Record;
+import org.jooq.SQLDialect;
+import org.jooq.UpdateSetFirstStep;
+import org.jooq.UpdateSetMoreStep;
+import org.jooq.impl.DSL;
+import org.junit.After;
+import org.junit.Before;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.sql.DataSource;
 import java.io.File;
 import java.io.FileReader;
 import java.sql.Connection;
@@ -13,38 +38,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
-
-import javax.sql.DataSource;
-
-import jet.isur.nsi.api.data.DictRow;
-import jet.isur.nsi.api.data.DictRowBuilder;
-import jet.isur.nsi.api.data.NsiConfig;
-import jet.isur.nsi.api.data.NsiConfigAttr;
-import jet.isur.nsi.api.data.NsiConfigDict;
-import jet.isur.nsi.api.data.NsiConfigField;
-import jet.isur.nsi.api.data.NsiQuery;
-import jet.isur.nsi.api.data.NsiQueryAttr;
-import jet.isur.nsi.api.model.DictRowAttr;
-import jet.isur.nsi.api.model.MetaAttrType;
-import jet.isur.nsi.common.data.DictDependencyGraph;
-import jet.isur.nsi.common.sql.DefaultSqlDao;
-import jet.isur.nsi.common.sql.DefaultSqlGen;
-import jet.isur.nsi.testkit.utils.DaoUtils;
-
-import org.jooq.DeleteWhereStep;
-import org.jooq.Field;
-import org.jooq.Record;
-import org.jooq.SQLDialect;
-import org.jooq.UpdateSetFirstStep;
-import org.jooq.UpdateSetMoreStep;
-import org.jooq.conf.RenderNameStyle;
-import org.jooq.conf.Settings;
-import org.jooq.impl.DSL;
-import org.jooq.impl.SQLDataType;
-import org.junit.After;
-import org.junit.Before;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public class BaseSqlTest {
 
@@ -217,7 +210,11 @@ public class BaseSqlTest {
         for (NsiQueryAttr attr : query.getAttrs()) {
            result.attr(attr.getAttr().getName(), (String) null);
         }
-        return result.deleteMarkAttr(false);
+
+        if(result.getDict().getDeleteMarkAttr() != null) {
+            return result.deleteMarkAttr(false);
+        }
+        return result;
     }
 
     protected NsiConfigDict dict(String dictName) {
